WebApp/requirements.txt
Flask==2.1.1
Flask_Cors==3.0.10
keras==2.8.0
numpy==1.22.3
Pillow==9.1.0
tensorflow==2.8.0
gunicorn